ok3

+++ src/main/java/hsf302/f25/s3/firstexam/controller/ProductController.java
@@ -1,6 +1,8 @@
 package hsf302.f25.s3.firstexam.controller;
 
+import hsf302.f25.s3.firstexam.entity.Account;
 import hsf302.f25.s3.firstexam.entity.Product;
+import hsf302.f25.s3.firstexam.service.AccountService;
 import hsf302.f25.s3.firstexam.service.CategoryService;
 import hsf302.f25.s3.firstexam.service.ProductService;
 import jakarta.servlet.http.HttpSession;
@@ -17,20 +19,33 @@ import org.springframework.web.bind.annotation.*;
 public class ProductController {
     private final ProductService productService;
     private final CategoryService categoryService;
+    private final AccountService accountService;
     @GetMapping
     public String getAllProducts(Model model, HttpSession session) {
+        Account account = (Account) session.getAttribute("account");
+        if (account == null || !accountService.isStaff(account) || !accountService.isAdmin(account)) {
+            return  "redirect:/login";
+        }
         model.addAttribute("products", productService.findAll());
         return "product";
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteProduct(@PathVariable Integer id) {
+    public String deleteProduct(@PathVariable Integer id, HttpSession session) {
+        Account account = (Account) session.getAttribute("account");
+        if (account == null || !accountService.isAdmin(account)) {
+            return  "redirect:/login";
+        }
         productService.deleteById(id);
         return "redirect:/products";
     }
 
     @GetMapping("/update/{id}")
-    public String updateProduct(@PathVariable Integer id, Model model) {
+    public String updateProduct(@PathVariable Integer id, Model model, HttpSession session) {
+        Account account = (Account) session.getAttribute("account");
+        if (account == null || !accountService.isAdmin(account)) {
+            return  "redirect:/login";
+        }
         model.addAttribute("products", productService.findById(id));
         model.addAttribute("categories", categoryService.findAll());
         model.addAttribute("formMode", "update");
@@ -38,7 +53,11 @@ public class ProductController {
     }
 
     @GetMapping("/create")
-    public String createProduct(Model model) {
+    public String createProduct(Model model, HttpSession session) {
+        Account account = (Account) session.getAttribute("account");
+        if (account == null || !accountService.isAdmin(account)) {
+            return  "redirect:/login";
+        }
         model.addAttribute("products", new Product());
         model.addAttribute("categories", categoryService.findAll());
         model.addAttribute("formMode", "create");
@@ -56,5 +75,16 @@ public class ProductController {
         }
         productService.save(product);
         return "redirect:/products";
+    }
+
+    @GetMapping("/top-stock")
+    public String getTopStock(Model model, HttpSession session){
+        Account account = (Account) session.getAttribute("account");
+        if(account == null || accountService.isAdmin(account)){
+            return "redirect:/login";
+        }
+
+        model.addAttribute("topProducts", productService.getTop3ProductsByStockPerCategory());
+        return "top-stock";
     }
 }
